+++ android/app/src/main/java/com/onair/hearit/presentation/main/DrawerClickListener.kt
@@ -1,4 +1,4 @@
-package com.onair.hearit.presentation
+package com.onair.hearit.presentation.main
 
 interface DrawerClickListener {
     fun openDrawer()
