+++ android/app/src/main/res/layout/item_searched_hearit.xml
@@ -6,7 +6,7 @@
     <data>
 
         <variable
-            name="item"
+            name="searchedHearit"
             type="com.onair.hearit.domain.model.SearchedHearit" />
 
         <variable
@@ -19,7 +19,7 @@
         android:layout_height="wrap_content"
         android:layout_marginBottom="16dp"
         android:background="@drawable/bg_gray1_radius_8dp"
-        android:onClick="@{() -> clickListener.onClick(item.id)}"
+        android:onClick="@{() -> clickListener.onClick(searchedHearit.id)}"
         android:paddingVertical="20dp">
 
         <TextView
@@ -31,7 +31,7 @@
             android:layout_marginEnd="8dp"
             android:ellipsize="end"
             android:maxLines="2"
-            android:text="@{item.title}"
+            android:text="@{searchedHearit.title}"
             app:layout_constraintEnd_toStartOf="@id/iv_searched_hearit"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
@@ -58,7 +58,7 @@
             android:textSize="12sp"
             app:layout_constraintStart_toStartOf="@id/tv_searched_hearit_title"
             app:layout_constraintTop_toBottomOf="@id/tv_searched_hearit_title"
-            tools:text="#kotlin  #Intent  #안드로이드" />
+            tools:text="# 바이브 코딩  # 컨텍스트 스위칭  # LLM" />
 
         <TextView
             android:id="@+id/tv_searched_hearit_time"
@@ -70,7 +70,7 @@
             android:textSize="12sp"
             app:layout_constraintEnd_toStartOf="@id/iv_searched_hearit"
             app:layout_constraintTop_toTopOf="@id/tv_keywords"
-            app:playTimeFormatted="@{item.playTime}"
+            app:playTimeFormatted="@{searchedHearit.playTime}"
             tools:text="08:21" />
 
         <ProgressBar
@@ -85,7 +85,7 @@
             app:layout_constraintEnd_toEndOf="@id/tv_searched_hearit_time"
             app:layout_constraintStart_toStartOf="@id/tv_searched_hearit_title"
             app:layout_constraintTop_toBottomOf="@id/tv_keywords"
-            app:progressRatio="@{item}" />
+            app:progressRatio="@{searchedHearit}" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
